requirements.txt
configobj~=5.0.9
pathlib~=1.0.1
joblib~=1.3.2
matplotlib~=3.10.0
xgboost~=2.1.4
scikit-learn~=1.6.1
Faker~=36.1.1
networkx~=3.2.1
numpy~=2.2.3
plotly~=6.0.0
seaborn~=0.13.2
torchviz~=0.0.3
tqdm~=4.66.6
requests~=2.32.3
DateTime~=5.5
sentence-transformers~=3.4.1
colorlog~=6.9.0
safetensors~=0.5.2
WMI~=1.5.1
prettytable~=3.14.0
pandas~=2.2.2
scapy~=2.5.0
psutil~=6.1.1